Source/Python/Services/Service3/requirements.txt
click==7.1.2
debugpy==1.2.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
